splitting out choiceFlag to its own file, formatted using goimport, altered pipelines.Contains to be private pipelines.contains
